Revert "Jdp230401 1"
