+++ src/main/java/com/kodilla/ecommercee/CartController.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee;
+
+public class CartController {
+}
