+++ src/main/java/com/kodilla/ecommercee/GroupController.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee;
+
+public class GroupController {
+}
