+++ src/main/java/com/kodilla/ecommercee/ProductController.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee;
+
+public class ProductController {
+}
