+++ src/main/java/com/kodilla/ecommercee/UserController.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee;
+
+public class UserController {
+}
