--- src/main/java/com/kodilla/ecommercee/controller/CartController.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.controller;
-
-public class CartController {
-}
