--- src/main/java/com/kodilla/ecommercee/controller/GroupController.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.controller;
-
-public class GroupController {
-}
