--- src/main/java/com/kodilla/ecommercee/controller/OrderController.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.controller;
-
-public class OrderController {
-}
