--- src/main/java/com/kodilla/ecommercee/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee.controller;
-
-public class UserController {
-}
